requirements.txt
requests==2.26.0
opencv-python==4.6.0
python-telegram-bot==13.11
beautifulsoup4==4.11.1
json==2.0.9
pillow==9.0.1